Implement an automatic bypass functionality in the generic client, with the following API:
  - setBypassEnabled
  - getBypassEnabled, both calls available from all threads. Using a raw atomics so no added memory barriers.

Implement "soft-bypass" support in VST2 and AU. Untestested since I don't find a VST2 host that uses it.
For AU bypass has just moved to the generic client, this should fixes some MIDI bugs in plugin with MIDI input + bypass.

In the future, a real "soft bypass" will be implementable internally with a specially-named parameter.