Delayline now has duplicated write pointers, in order to make any delayline read always contiguous. This speed-up reading taps since the same pointer can be accessed all the time. Panagement 2 gets 6.5% faster thanks to this.
For all purposes, the new gets about -120 dB abs difference with the old => very close to being identical.
API hasn't changed, so this is marked as non-breaking.
Sound got enhanced because index are decomposed to fractional part in double precision, which improve things on very long dealy line (~20000 samples and above).