In dplug:gui, introduce user pixel space and logical pixel space.
The host and the window talk in "logical pixels".
The UI (widgets) talk in "user pixels".

GUIGraphics.doResize now accepts any incoming logical size.
The single entry point for UI resizing is now IWindowListener.onResized
The final buffer is cropped and offset in order to best fit the user area into the logical area (_userArea rect).

Note that this is unrelated to DPI, which introduce separation between physical pixels and logical pixels.
This will also help on iPad which give you a surface with any size.