sysconf is a Posix utility
It is defined in POSIX.1 and first appeared in 4.4BSD. While it is true that _SC_NPROCESSORS_ONLN is an extension, many systems provide it.
As such, it should not be marked version(linux). It should be marked version(Posix).
Unbreaks the build on non-Linux systems (OpenBSD, in my case).