Proper handling of tail size (Issue #622).
Possibly breaking, risk is unknown: default tail size is now +infinity, which is the correct one for synths and effects generating sound.
It should be safer for everyone (people generally give too short a tail size, or forget about it). The previous default was 2 seconds, which didn't mean anything.
It is possible, though unlikely, that we get a few breakage in AU and VST3 hosts that can't handle infinite tail size.